# Improve caching performance by adjusting default call pattern (take 2)